Add welcome message function to dom updates file
